+++ launchdarkly-android-client-sdk/src/main/java/com/launchdarkly/android/UserManager.java
@@ -1,212 +1,9 @@
 package com.launchdarkly.android;
 
-import android.app.Application;
-import android.content.SharedPreferences;
-import android.support.annotation.NonNull;
-import android.support.annotation.VisibleForTesting;
-import android.util.Base64;
-
-import com.google.gson.JsonObject;
-import com.launchdarkly.android.flagstore.Flag;
-import com.launchdarkly.android.flagstore.FlagStore;
-import com.launchdarkly.android.flagstore.FlagStoreManager;
-import com.launchdarkly.android.flagstore.sharedprefs.SharedPrefsFlagStoreFactory;
-import com.launchdarkly.android.flagstore.sharedprefs.SharedPrefsFlagStoreManager;
-import com.launchdarkly.android.gson.GsonCache;
-import com.launchdarkly.android.response.DeleteFlagResponse;
-import com.launchdarkly.android.response.FlagsResponse;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.concurrent.ExecutorService;
-
-import timber.log.Timber;
-
-/**
- * Persists and retrieves feature flag values for different {@link LDUser}s.
- * Also enables realtime updates via registering a {@link FeatureFlagChangeListener}
- * with a feature flag.
- */
-class UserManager {
-
-    private final FeatureFlagFetcher fetcher;
-
-    private final Application application;
-    private final FlagStoreManager flagStoreManager;
-    private final SummaryEventSharedPreferences summaryEventSharedPreferences;
-    private final String environmentName;
-
-    private LDUser currentUser;
-
-    private final ExecutorService executor;
-
-    static synchronized UserManager newInstance(Application application, FeatureFlagFetcher fetcher, String environmentName, String mobileKey) {
-        return new UserManager(application, fetcher, environmentName, mobileKey);
-    }
-
-    UserManager(Application application, FeatureFlagFetcher fetcher, String environmentName, String mobileKey) {
-        this.application = application;
-        this.fetcher = fetcher;
-        this.flagStoreManager = new SharedPrefsFlagStoreManager(application, mobileKey, new SharedPrefsFlagStoreFactory(application));
-        this.summaryEventSharedPreferences = new UserSummaryEventSharedPreferences(application, LDConfig.SHARED_PREFS_BASE_KEY + mobileKey + "-summaryevents");
-        this.environmentName = environmentName;
-
-        executor = new BackgroundThreadExecutor().newFixedThreadPool(1);
-    }
-
-    LDUser getCurrentUser() {
-        return currentUser;
-    }
-
-    FlagStore getCurrentUserFlagStore() {
-        return flagStoreManager.getCurrentUserStore();
-    }
-
-    SummaryEventSharedPreferences getSummaryEventSharedPreferences() {
-        return summaryEventSharedPreferences;
-    }
-
-    /**
-     * Sets the current user. If there are more than MAX_USERS stored in shared preferences,
-     * the oldest one is deleted.
-     *
-     * @param user The user to switch to.
-     */
-    void setCurrentUser(final LDUser user) {
-        String userBase64 = user.getAsUrlSafeBase64();
-        Timber.d("Setting current user to: [%s] [%s]", userBase64, userBase64ToJson(userBase64));
-        currentUser = user;
-        flagStoreManager.switchToUser(user.getSharedPrefsKey());
-    }
-
-    void updateCurrentUser(final Util.ResultCallback<Void> onCompleteListener) {
-        fetcher.fetch(currentUser,
-                new Util.ResultCallback<JsonObject>() {
-                    @Override
-                    public void onSuccess(JsonObject result) {
-                        saveFlagSettings(result, onCompleteListener);
-                    }
-
-                    @Override
-                    public void onError(Throwable e) {
-                        if (Util.isClientConnected(application, environmentName)) {
-                            Timber.e(e, "Error when attempting to set user: [%s] [%s]",
-                                    currentUser.getAsUrlSafeBase64(),
-                                    userBase64ToJson(currentUser.getAsUrlSafeBase64()));
-                        }
-                        onCompleteListener.onError(e);
-                    }
-                });
-    }
-
-    void registerListener(final String key, final FeatureFlagChangeListener listener) {
-        flagStoreManager.registerListener(key, listener);
-    }
-
-    void unregisterListener(String key, FeatureFlagChangeListener listener) {
-        flagStoreManager.unRegisterListener(key, listener);
-    }
-
-    void registerAllFlagsListener(@NonNull final LDAllFlagsListener listener) {
-        flagStoreManager.registerAllFlagsListener(listener);
-    }
-
-    void unregisterAllFlagsListener(@NonNull final LDAllFlagsListener listener) {
-        flagStoreManager.unregisterAllFlagsListener(listener);
-    }
-
-    /**
-     * Saves the flags param to {@link SharedPreferences} for the current user.
-     * Completely overwrites all values in the current user's {@link SharedPreferences} and
-     * saves those values to the active user, triggering any registered {@link FeatureFlagChangeListener}
-     * objects.
-     *
-     * @param flagsJson
-     */
-    @SuppressWarnings("JavaDoc")
-    private void saveFlagSettings(JsonObject flagsJson, Util.ResultCallback<Void> onCompleteListener) {
-        Timber.d("saveFlagSettings for user key: %s", currentUser.getKey());
-
-        try {
-            final List<Flag> flags = GsonCache.getGson().fromJson(flagsJson, FlagsResponse.class).getFlags();
-            flagStoreManager.getCurrentUserStore().clearAndApplyFlagUpdates(flags);
-            onCompleteListener.onSuccess(null);
-        } catch (Exception e) {
-            Timber.d("Invalid JsonObject for flagSettings: %s", flagsJson);
-            onCompleteListener.onError(new LDFailure("Invalid Json received from flags endpoint", e, LDFailure.FailureType.INVALID_RESPONSE_BODY));
-        }
-    }
-
-    private static String userBase64ToJson(String base64) {
-        return new String(Base64.decode(base64, Base64.URL_SAFE));
-    }
-
-    void deleteCurrentUserFlag(@NonNull final String json, final Util.ResultCallback<Void> onCompleteListener) {
-        try {
-            final DeleteFlagResponse deleteFlagResponse = GsonCache.getGson().fromJson(json, DeleteFlagResponse.class);
-            executor.submit(new Runnable() {
-                @Override
-                public void run() {
-                    if (deleteFlagResponse != null) {
-                        flagStoreManager.getCurrentUserStore().applyFlagUpdate(deleteFlagResponse);
-                        onCompleteListener.onSuccess(null);
-                    } else {
-                        Timber.d("Invalid DELETE payload: %s", json);
-                        onCompleteListener.onError(new LDFailure("Invalid DELETE payload",
-                                LDFailure.FailureType.INVALID_RESPONSE_BODY));
-                    }
-                }
-            });
-        } catch (Exception ex) {
-            Timber.d(ex, "Invalid DELETE payload: %s", json);
-            onCompleteListener.onError(new LDFailure("Invalid DELETE payload", ex,
-                    LDFailure.FailureType.INVALID_RESPONSE_BODY));
-        }
-    }
-
-    void putCurrentUserFlags(final String json, final Util.ResultCallback<Void> onCompleteListener) {
-        try {
-            final List<Flag> flags = GsonCache.getGson().fromJson(json, FlagsResponse.class).getFlags();
-            executor.submit(new Runnable() {
-                @Override
-                public void run() {
-                    Timber.d("PUT for user key: %s", currentUser.getKey());
-                    flagStoreManager.getCurrentUserStore().clearAndApplyFlagUpdates(flags);
-                    onCompleteListener.onSuccess(null);
-                }
-            });
-        } catch (Exception ex) {
-            Timber.d(ex, "Invalid PUT payload: %s", json);
-            onCompleteListener.onError(new LDFailure("Invalid PUT payload", ex,
-                    LDFailure.FailureType.INVALID_RESPONSE_BODY));
-        }
-    }
-
-    void patchCurrentUserFlags(@NonNull final String json, final Util.ResultCallback<Void> onCompleteListener) {
-        try {
-            final Flag flag = GsonCache.getGson().fromJson(json, Flag.class);
-            executor.submit(new Runnable() {
-                @Override
-                public void run() {
-                    if (flag != null) {
-                        flagStoreManager.getCurrentUserStore().applyFlagUpdate(flag);
-                        onCompleteListener.onSuccess(null);
-                    } else {
-                        Timber.d("Invalid PATCH payload: %s", json);
-                        onCompleteListener.onError(new LDFailure("Invalid PUT payload",
-                                LDFailure.FailureType.INVALID_RESPONSE_BODY));
-                    }
-                }
-            });
-        } catch (Exception ex) {
-            Timber.d(ex, "Invalid PATCH payload: %s", json);
-            onCompleteListener.onError(new LDFailure("Invalid PATCH payload", ex,
-                    LDFailure.FailureType.INVALID_RESPONSE_BODY));
-        }
-    }
-
-    @VisibleForTesting
-    Collection<FeatureFlagChangeListener> getListenersByKey(String key) {
-        return flagStoreManager.getListenersByKey(key);
-    }
+interface UserManager {
+    LDUser getCurrentUser();
+    void putCurrentUserFlags(String json, Util.ResultCallback<Void> onCompleteListener);
+    void patchCurrentUserFlags(String json, Util.ResultCallback<Void> onCompleteListener);
+    void deleteCurrentUserFlag(String json, Util.ResultCallback<Void> onCompleteListener);
+    void updateCurrentUser(Util.ResultCallback<Void> onCompleteListener);
 }
